nvmem: core: allow single and dynamic device ids
At the moment only DEVICE_ID_DYNAMIC is supported. Add support for
NVMEM_DEVID_NONE to allow static device names via of-aliases and to
align our core with Linux.

While on it fix the nvmem_register_cdev() by using the dev_name() which
honors the DEVICE_ID_* cases else we end up with different names for the
devfs and the device itself.

The dev_dbg() message can be dropped as well since the register_device()
can print this message too.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20250818-v2025-06-0-topic-nvmem-v2-7-b6d677f6c519@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>