of: overlay: make the pattern match function more generic
The current overlay mechanism can handle files only, so filepattern was
an obvious name.

Drop the "file" prefix to prepare the codebase to handle overlays
supplied via a FIT images as well.

Keep the backward compatibility by still providing the filepattern
filter and the global of.overlay.filepattern variable, but mark them as
deprecated.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20250820-v2024-05-0-topic-fit-overlay-v6-3-9d1b3bab2ee2@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>