scripts: rockchip: implement image signing
The header of a Rockchip signed images contains an RSA public key and
the signature generated with the corresponding RSA private key at the
end of the header.

Since barebox already creates the header for the Rockchip image, adding
the signature in the rkimage tool is very convenient and avoids the need
for a separate tool for signing.

Add the -k option to the rkimage tool to pass the path to a PEM file or
a PKCS#11 uri of the private RSA key that shall be used for signing.

Extend the barebox build system to set this parameter and pass
CONFIG_ARCH_ROCKCHIP_IMAGE_SIGN_KEY as key if
CONFIG_ARCH_ROCKCHIP_IMAGE_SIGNED is enabled. The option may be
overridden by an environment variable of the same name to allow an
easier integration into other build systems.

The generated and signed images can be verified by using the Rockchip
vendor tool [0] as:

	rk_sign_tool vb --idb <barebox-image>

The Rockchip vendor tool isn't used for signing the images. rk_sign_tool
is difficult to use in an automated environment, because it uses a
settings.ini file instead of command line parameters. Furthermore, the
integration of rk_sign_tool with an HSM or PKCS#11 is not documented.

As it is useful to be able to resign an existing Rockchip image with a
different key, extend the rkimage to accept Rockchip images as input and
skip the IDB creation in this case.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Link: https://lore.barebox.org/20250821-sign-rkimage-v1-1-910d74532724@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>