ARM: mmu: increase early page table size to 256K for now
With recent changes for W^X, the size of our early page tables have grown
in some cases, so they no longer fit into the 64K we reserve for them.

This has been observed in at least two ARM64 platforms:

  - i.MX8M with OP-TEE at start of RAM (on master)
  - Raspberry Pi Zero 2 without OP-TEE (on next)

It's very likely we are overly wasteful of space, but until that's fixed
properly, increase the size reserved for our early page tables.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Reviewed-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20250822140111.811702-1-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>