kbuild: move PBL_CODE_SIZE_ out of rule_pblb
Before Commit 21613ae35f ("kbuild: images/Makefile: fix broken if_changed usage")
setting of PBL_CODE_SIZE_$*, PBL_MEMORY_SIZE_$* and PBL_IMAGE_SIZE_$*
were part of the

$(obj)/%.pblb: $(obj)/%.pbl FORCE

rule which was always executed due to the FORCE. Beginning with this
commit setting of these variables are part of rule_pblb which is only
executed when the pblb file is (re-)generated. In case the pblb is still
up to date this results in the variables being unset which then fails
in later stages.

Move setting of the variables back to the pblb rule to make sure they
are always set.

Reproducer:

make multi_v8_defconfig && make && make

The second make will fail while calling the imx-image tool which gets
the -p option without the actual size as optarg.

Fixes: 21613ae35f ("kbuild: images/Makefile: fix broken if_changed usage")
Link: https://lore.barebox.org/20250825131806.2050398-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>