public-keys: store pointers in linker list
On X86 the linker puts the structs in a linker list at 8 byte
boundaries, but iterating over the list advances the pointer by the
actual struct size which might not be 8 byte aligned.

Instead of collecting the actual structs in the linker list, store
pointers to the structs in the list like we do in other places already.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>