Kbuild: do not build dtbs concurrently
commit 3c862a83f019 ("kbuild: restrict dtbs target to enabled DTs by
default") added the dtbs target as a dependency to the images target.
In the dtbs target we call into Kbuild for building the device trees.
The device trees are also built as part of the regular build in
parallel, so we end up running Kbuild for the same directory twice.

This sometimes results in failed builds with messages like this one
(done with make V=2):

  AS [P]  arch/arm/dts/bcm2711-rpi-400.dtb.pbl.o - due to target missing
  AS [P]  arch/arm/dts/bcm2711-rpi-400.dtb.pbl.o - due to target missing
fixdep: error opening file: arch/arm/dts/.bcm2711-rpi-400.dtb.pbl.o.d: No such file or directory

Fix this by removing the dtbs target from the images dependencies. We
then would no longer build the dtbs-list, so build that unconditionally
once we descend into a directory containing device trees. The dtbs-list
is built with always-y which should make sure it is ready once we need
it in imagages/

Fixes: 3c862a83f019 ("kbuild: restrict dtbs target to enabled DTs by default")
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250828093135.1959085-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>