vsprintf: drop printf hexstr fuzzer
This fuzzer was an early proof of concept as can be seen by the
useless initialized variable and its name is misleading:
It only tests the hexstr feature.

Unlike other tests, it also spams stdout. Let's just drop it for now and
perhaps bring it back in future in a more complete form.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250901115054.2514841-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>