string: add strtokv
Add an implementation of libc's standard strtok(3), which is useful
for tokenizing strings.

Also, add a version that will collect all tokens from a string into an
array, which is useful in situations where you need to know how many
tokens there are, and when a token's relative position in the order is
significant.

Signed-off-by: Tobias Waldekranz <tobias@waldekranz.com>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250909131843.2260573-2-tobias@waldekranz.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>