MIPS: qemu-malta_defconfig: Use largest possible relocation table
With the upcoming device-mapper selftest, enabling CONFIG_DISK and
CONFIG_DM_BLK pushed the resulting relocation table over the 0x8000
default.

Given that the target hardware is QEMU, RAM is unlikely to be a scarce
resource. Therefore, go for the maximum allowed to avoid being back
here again when the next thing is added.

Signed-off-by: Tobias Waldekranz <tobias@waldekranz.com>
Link: https://lore.barebox.org/20250909131843.2260573-6-tobias@waldekranz.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>