drivers: dma: refactor: rename dma_ops to dma_device_ops.
This patch is an esthetic cleanup, rename the dma_ops to dma_device_ops
to be sure that the struct in drivers/dma is different from one used in
arch/riscv/cpu. I accidentally faced the issue telling that the struct is
defined in the two places but it is not supposed to happen since the
the driver is never selected from riscv.

Signed-off-by: Chali Anis <chalianis1@gmail.com>
Link: https://lore.barebox.org/20250914193249.23414-1-chalianis1@gmail.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>