ARM: rockchip: cleanup iram handling
struct rockchip_scratch_space has grown and gained additional fields.
While the scratch space for the iram is still at the start of the
rockchip_scratch_space, it is actually just one field in the scratch
space.

Explicitly store the iram to the iram field, and copy only the size of
the iram field instead of the entire scratch space.

Explicitly pass the iram field when the iram should be parsed.

Rename the field from irom to iram.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Link: https://lore.barebox.org/20250916-rockchip-cleanup-iram-v1-1-e6797d08f8dc@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>