ARM: configs: add virt32_secure_defconfig
The security policy support does not allow for incomplete configs and
thus sconfig files must be refreshed when config options they depend on
changes. This means that a security profile that's up-to-date with
respect to one .config is often outdated with respect to another.

To allow easy development and experimentation, let's make 32-bit ARM
Qemu Virt our reference platform and add a new config for it.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250917-security-policies-v2-14-f30769a3ff51@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>