security: shell: add policy for executing the shell
Executing shell scripts can be dangerous in secure environments, so add
a security policy for it. While shell scripts can be executed securely
if made sure that no scripts from unknown sources are executed,
executing an interactive shell for sure is not desired in secure
environments, so offer two options: One for disabling the shell entirely
and one for disabling interactive shells.

Link: https://lore.barebox.org/20250917-security-policies-v2-21-f30769a3ff51@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>