doc: bootchooser: use 0 as base for system targets
Most barebox-state configurations in upstream define the bootchooser
targets as system0 and system1. The RAUC bootchooser documentation uses
system0 and system1, too. This is different from the bootchooser
documentation in barebox, which uses system1 and system2 in its
examples.

Change the barebox documentation to be consistent with what most users
will probably find on their respective systems.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Link: https://lore.barebox.org/20250926-bootchooser-doc-v1-1-d07fb39ffacd@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>