MAKEALL: query CONFIG_64BIT before make instead of CONFIG_ARM64 after
The current logic inside MAKEALL is flawed: We source an ARM64 config
with CROSS_COMPILE=${CROSS_COMPILE_arm} and hope that CONFIG_ARM64
results, so we can grep for it in the .config despite it missing in the
defconfig.

Replace this by looking for CONFIG_64BIT instead in the defconfig
directly. This symbol seems to be there for all ARM64 configs we have
and if this ceases, CI will let us know.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250930085401.2133061-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>