Revert "MAKEALL: query CONFIG_64BIT before make instead of CONFIG_ARM64 after"
The commit being reverted here assumed that all defconfigs have a
corresponding file, but this doesn't hold true for generated configs
like multi_v8_efi_defconfig.

The issue that was meant to be addressed disappeared after rerunning
the test, so it might have been just a fluke (or triggered by shuffle,
although I was unable to reproduce with the same seed).

Until better understood, reinstate the old code.

Fixes: bdf0a4da1d28 ("MAKEALL: query CONFIG_64BIT before make instead of CONFIG_ARM64 after")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250930103822.2395670-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>