ARM: dts: am62lx: fix secondary core startup
Disable the rti1 watchdog which fixes startup of the secondary CPU core
under Linux.

Enabling the rti1 watchdog leads to enabling its power domain,
<&scmi_pds 61> in this case. With this starting the secondary core fails
under Linux:

[    5.095816] CPU1: failed to come online
[    5.095832] CPU1: failed in unknown state : 0x0
[    5.095892] smp: Brought up 1 node, 1 CPU
[    5.095898] SMP: Total of 1 processors activated.

I don't know what's exactly going on here, but it seems enabling the
power domain either indirectly starts the secondary core in barebox
already or it causes some weird power state for the secondary CPU from
which the Tf-A can't recover later.

The device node is the same as under Linux later, but in Linux the rti1
watchdog is configured after the secondary CPU has been started.

Link: https://lore.barebox.org/20251007085117.941408-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>