FIT: compare key name hint with streq_ptr for clarity
When given a key name hint, we first attempt verification using the key
it describes and fall through to verifying with all other keys.

The code implementing this is not necessarily intuitive and strcmp()
makes it harder to understand.

As we check for NULL pointer anyway, let's use streq_ptr, which returns
true instead of 0 on equality.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Reviewed-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20251010054939.3067504-1-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>