MAKEALL: print --shuffle seed used for initial make *_defconfig
We run make *_defconfig once initially, so that we can get a .config to
grep for CONFIG_ARM64 inside. The grepping occasionally[1] fails in CI:

  grep: build-arm-multi_v8_defconfig/.config: No such file or directory

This might be because of the shuffling of the dependencies, but because
make itself doesn't fail, we don't get the shuffle seed in the log.

Let's collect it and print it to help track down this issue.

[1]: https://github.com/barebox/barebox/actions/runs/18654962634/job/53181695019

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20251021075742.4016898-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>