Documentation: gen_commands.py: escape special characters
commands/global.c has following line:

  BAREBOX_CMD_HELP_TEXT("Without options, print all global and env
                        (prefixed with *) variables.")

which triggers a warning as the * has a special meaning in ReST and it
is never terminated.

On the other hand, we have escapes for double quotes, which should be
stripped:

  BAREBOX_CMD_HELP_TEXT("- \"bootchooser\": boot with barebox bootchooser")

Fix these and other issues by processing escapes for all strings we extract,
except for those in literal blocks.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20251024090708.2476109-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>