Documentation: gen_commands.py: use literal block for help text
The help text can have lists, like in the boot command[1] and printing
this as normal text messes up the formatting.
Therefore, let's just print it as a code block.

This doesn't solve interleaved output like fiptool's, but it makes e.g.
the boot help text or menu actually readable.

[1]: https://www.barebox.org/doc/latest/commands/boot/boot.html

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20251024090708.2476109-4-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>