kbuild: add make check target
The barebox test suite is invoked by running pytest, but for many
projects running make check is what is usually expected.

Add a make check target that calls pytest to accommodate that.
In future, we may add extra checks to test that pytest is correctly
installed to make it easier to use.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20251027080832.1989341-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>