pmdomain: look up pmdomain even if not have_genpd_providers
have_genpd_providers is a pre-deep-probe artifact, where SoC support
would promise that a power domain driver will be registered.

In a deep probe system, we can just force the probe in
__genpd_dev_pm_attach.

Therefore, move the check for have_genpd_providers, until after the
lookup of power domain drivers.

This will be further refined in follow-up commits.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20251027192938.2586566-2-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>