commands: dmesg: clear logbuffer fully if not requested otherwise
Previously, `dmesg -c` would clear the logbuffer except for the last 10
lines. This comes with a fair amount of surprise given linux `dmesg -c`
clears the whole logbuffer.

Clear the complete logbuffer given `dmesg -c` but with an optional
argument `dmesg -c [<num>]` to keep <num> lines.

While at it, add documentation for log_print(), and switch it from
ctrlc() to ctrlc_non_interruptible(). This ensures that the complete
execution of the dmesg command has no resched() point so that no other
thread can intermittently modify the logbuffer, and that no message
could arrive from another thread during log_print() and be interleaved
with the dmesg output.

If loglevels are selected using -l or -p in addition to clearing being
selected using -c, then messages of any loglevel are cleared.

Signed-off-by: Jonas Rebmann <jre@pengutronix.de>
Link: https://lore.barebox.org/20251029-dmesg-v2-1-c54c7daef91b@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>