gitignore: ignore /build, /oe-logs, /oe-workdir
Documentation such as Documentation/boards/kvx.rst and tooling such as
conftest.py suggests /build is a good choice for build output directory.
The build system does set up a .gitignore in the build directory but
this has no effect if /build is a symlink to outside the tree.

Similarly, YOCTOs "devtool" sets up the symlinks /oe-logs and
/oe-workdir to point to the respective paths in yoctos build tree.

Include those in .gitignore.

Signed-off-by: Jonas Rebmann <jre@pengutronix.de>
Link: https://lore.barebox.org/20251030-gitignore-build-v2-1-8f4fcd824921@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>