usb: dwc3: port Linux i.MX8MP glue driver
Commit e213627bbe ("usb: dwc3: of-simple: add i.MX8MP compatible")
added i.MX8MP support by matching the i.MX8MP glue in dwc3-of-simple.

This is unlike what the kernel was doing, because "Linux uses this wrapper
for wakeup handling, but we don't need this functionality in Barebox".

Looking at Linux v6.18-rc3, this is not accurate anymore:
The upstream i.MX8MP glue driver handles a number of device tree properties
that we do not handle in barebox:

  * fsl,permanently-attached
  * fsl,disable-port-power-control
  * fsl,over-current-active-low
  * fsl,power-active-low

The Linux driver also sets two software properties:

  * xhci-missing-cas-quirk
  * xhci-skip-phy-init-quirk

But both are unneeded for barebox as the former only matters when
resuming from runtime suspend and the latter is the only behavior we
currently have in barebox, because the XHCI driver itself doesn't yet
use the PHY API.

Thus import the Linux driver without the software property bits to add
support for the OF bindings.

Reported-by: Marc Kleine-Budde <mkl@pengutronix.de>
Reported-by: Holger Assmann <h.assmann@pengutronix.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Tested-by: Thorsten Scherer <t.scherer@eckelmann.de>
Link: https://lore.barebox.org/20251103151935.2368094-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>